Split client in several files
